rename json_manager
