+++ dqt/graph.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING, Any
 from types import NoneType
 
 from dqt.dqt_json import DQTJSON
-from dqt.json_manager import JsonManager
+from dqt.settings_manager import SettingsManager
 from dqt.ui_utils import err, confirm
 
 try:
@@ -204,11 +204,11 @@ class Graph:
             TypeError: Incorrect type
         """
 
-        if JsonManager.settings is None:
+        if SettingsManager.settings is None:
             print("JSON manager has not been loaded, cannot configure graph")
             return
         
-        settings: dict[str, dict[str, Any]] = JsonManager.settings
+        settings: dict[str, dict[str, Any]] = SettingsManager.settings
 
         for setting_name, value_dict in settings['graph'].items():
             if setting_name not in self._SETTING_KEYS:
